+++ src/js/race.js
@@ -1,6 +1,6 @@
-import { CAR_RACE, WINNER_MESSAGE } from './constant.js';
+import { CAR_RACE } from './constant.js';
 import { renderForward } from './ui/dom.js';
-import { getRandom } from './utils/util.js';
+import { getRandom } from './utils/random.js';
 
 export const startRace = (cars, times) => {
   return new Promise((resolve) => {
@@ -11,7 +11,7 @@ export const startRace = (cars, times) => {
       cars.map((car) => {
         const forwardedStep = getRandom(CAR_RACE.MIN, CAR_RACE.MAX);
         if (forwardedStep >= CAR_RACE.FORWARD) {
-          car.forwardStep(forwardedStep);
+          car.forwardStep();
           renderForward(car.name);
         }
       });
@@ -26,7 +26,6 @@ export const startRace = (cars, times) => {
 
 export const getWinner = (raceResult) => {
   const maxStep = Math.max(...raceResult.map((car) => car.step));
-  setTimeout(() => alert(WINNER_MESSAGE), 2000);
   return raceResult
     .filter((car) => car.step === maxStep)
     .map((car) => car.name);
