+++ src/js/utils/fp.js
@@ -1,0 +1,6 @@
+export const tab =
+  (value) =>
+  (...callbacks) => {
+    callbacks.forEach((callback) => callback(value));
+    return value;
+  };
