+++ src/js/utils/random.js
@@ -1,0 +1,3 @@
+export const getRandom = (min, max) => {
+  return Math.random() * (max - min) + min;
+};
